requirements.txt
Flask
numpy
Pillow
gdown
requests
fastapi
uvicorn[standard]

-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cpu 
torchvision==0.9.1+cpu 
torchaudio==0.8.1 

grad-cam

